+++ CHANGELOG.md
@@ -26,6 +26,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Fixed context menu opening error that could occur in cohort builder when datasets are not configured properly (e.g. have too many [IsExtractionIdentifier] columns).
 - Fixed alias changes not showing up as 'Differences' in edit dataeset extraction user interface
 - Fixed bugs in using GoTo menu of document tabs after a Refresh
+- Fixed ALTER context sub menu of TableInfo when Server property is null (or other fundamental connection details cannot be resolved).
 
 ## [4.1.3] - 2020-06-15